FotsalManager_final/requirements.txt
django
Pillow
django-htmx
django-htmx-refresh
django-allauth
django-crispy-forms
django-debug-toolbar
crispy-bootstrap5
folium
geocoder
easy_thumbnails
django-filer
django-mptt